+++ web/src/app.js
@@ -6,9 +6,9 @@ import thunkMiddleware from 'redux-thunk';
 import logger from 'redux-logger';
 import { createStore, applyMiddleware, combineReducers } from 'redux';
 import { search, getItem, setSearchForm, setSelectedPath, setSelectedItem,  setAdjacentItemIds, setLastQuery, toggleView } from './actions';
-import SearchBox from './search-box';
+import SearchBox from './components/search-box';
 import Box from '@material-ui/core/Box';
-import ItemBox from './item-box';
+import ItemBox from './components/item-box';
 import { renderRoutes } from 'react-router-config';
 import { createMuiTheme } from '@material-ui/core/styles';
 import * as colors from '@material-ui/core/colors';
