+++ web/src/components/content-box.js
@@ -1,6 +1,6 @@
 import React from 'react';
 import { renderRoutes } from 'react-router-config';
-import { routes } from './app';
+import { routes } from '../app';
 import Box from '@material-ui/core/Box';
 
 const ContentBox = props => {
